Rename TableDefinition to StandardTableDefinition and BaseTableDefinition to TableDefinition
