+++ README.md
@@ -127,12 +127,12 @@ must [supply credentials](#authentication) and a project ID if running this snip
 ```java
 import com.google.gcloud.bigquery.BigQuery;
 import com.google.gcloud.bigquery.BigQueryOptions;
-import com.google.gcloud.bigquery.TableDefinition;
 import com.google.gcloud.bigquery.Field;
 import com.google.gcloud.bigquery.JobStatus;
 import com.google.gcloud.bigquery.JobInfo;
 import com.google.gcloud.bigquery.LoadJobConfiguration;
 import com.google.gcloud.bigquery.Schema;
+import com.google.gcloud.bigquery.StandardTableDefinition;
 import com.google.gcloud.bigquery.TableId;
 import com.google.gcloud.bigquery.TableInfo;
 
@@ -143,7 +143,7 @@ if (info == null) {
   System.out.println("Creating table " + tableId);
   Field integerField = Field.of("fieldName", Field.Type.integer());
   Schema schema = Schema.of(integerField);
-  bigquery.create(TableInfo.of(tableId, TableDefinition.of(schema)));
+  bigquery.create(TableInfo.of(tableId, StandardTableDefinition.of(schema)));
 } else {
   System.out.println("Loading data into table " + tableId);
   LoadJobConfiguration configuration = LoadJobConfiguration.of(tableId, "gs://bucket/path");
