+++ gcloud-java-bigquery/src/test/java/com/google/gcloud/bigquery/TableTest.java
@@ -55,7 +55,8 @@ public class TableTest {
   private static final JobInfo EXTRACT_JOB_INFO =
       JobInfo.of(ExtractJobConfiguration.of(TABLE_ID1, ImmutableList.of("URI"), "CSV"));
   private static final Field FIELD = Field.of("FieldName", Field.Type.integer());
-  private static final BaseTableDefinition TABLE_DEFINITION = TableDefinition.of(Schema.of(FIELD));
+  private static final TableDefinition TABLE_DEFINITION =
+      StandardTableDefinition.of(Schema.of(FIELD));
   private static final TableInfo TABLE_INFO = TableInfo.of(TABLE_ID1, TABLE_DEFINITION);
   private static final List<RowToInsert> ROWS_TO_INSERT = ImmutableList.of(
       RowToInsert.of("id1", ImmutableMap.<String, Object>of("key", "val1")),
